rename module containing context managers
